further refcatoring: removing redundant jsx tags, renaming componentsh
