--- src/components/Layout/Layout.js
@@ -1,17 +1,0 @@
-import React, { Component } from 'react';
-import Toolbar from '../../components/Navigation/Toolbar/Toolbar';
-import IntroPage from './IntroPage';
-
-class Layout extends Component {
-
-    render () {
-        return (
-            <React.Fragment>
-                <Toolbar/>
-                <IntroPage />
-            </React.Fragment>
-        )
-    }
-}
-
-export default Layout;
