--- src/components/Layout/SocialMediaIcons/SocialMediaIcons.css
@@ -1,7 +1,0 @@
-.SocialMediaIcons {
-  padding-top: 240px;
-}
-
-div.SocialMediaIcons > span {
-  margin-left: 25px;
-}
