--- src/components/Navigation/Footer/Footer.css
@@ -1,10 +1,0 @@
-.Footer > a {
-  background-color: rgba(26,26,26,.75);
-  height: 450px;
-  text-align: center;
-}
-
-.Footer {
-  background-color: #262628;
-  color: #ffffff;
-}
