+++ src/components/Navigation/NavigationItems.js
@@ -1,12 +1,11 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import classes from './NavigationItems.css'
-import IntroPage from '../../Layout/IntroPage'
-import Experience from '../../Experience/Experience'
-import AboutMe from '../../AboutMe';
+import IntroPage from '../Layout/IntroPage'
+import Experience from '../Experience/Experience'
+import AboutMe from '../AboutMe';
 
 const NavigationItems = () => (
-    <ul className={`${classes.NavigationItems} navFooter`}>
+    <ul className='navigationItems'>
         <Link component={IntroPage} to={"/"}>Home</Link>     
         <Link component={Experience} to="/experience">Experience</Link>
         <Link component={AboutMe} to="/aboutme">About</Link>
