--- src/components/Navigation/NavigationItems/NavigationItems.css
@@ -1,48 +1,0 @@
-.NavigationItems {
-    padding: 0;
-    list-style: none;
-    display: flex;
-    /* flex-direction: row-reverse; */
-    height: 100%;
-    /* justify-content: flex-end; */
-    font-family: Raleway, monospace; 
-}
-
-.NavigationItems a {
-    color: #fff;
-    text-decoration: none;
-    width: 100%;
-    box-sizing: border-box;
-}
-
-@media (min-width: 500px) {
-    .NavigationItems {
-        flex-flow: row;
-    }
-}
-
-@media (min-width: 500px) {
-    .NavigationItems {
-        margin: 0;
-        display: flex;
-        height: 100%;
-        width: auto;
-        align-items: center;
-    }
-    
-    .NavigationItems a {
-        color: #fff;
-        height: 100%;
-        padding: 16px 10px;
-        border-bottom: 4px solid transparent;
-    }
-    
-    .NavigationItems a:hover,
-    .NavigationItems a:active,
-    .NavigationItems a.active {
-        background-color: rgba(5, 5, 5, 0.625);
-        border-bottom: 4px solid #40A4C8;
-        color: #fff;
-        padding-bottom: 2px;
-    }
-}
